requirements.txt
python-dotenv==0.12.0
PyYAML==5.3
selenium==3.141.0
urllib3==1.25.8
